+++ tests/test_python_interface.py
@@ -1,9 +1,10 @@
 import inspect
 import os
-from datetime import datetime
+from datetime import datetime, timedelta
 from pathlib import Path
 
 from copernicusmarine import (
+    core_functions,
     describe,
     get,
     login,
@@ -206,3 +207,15 @@ class TestPythonInterface:
         assert dataset.depth.attrs["positive"] == "down"
         assert dataset.depth.attrs["standard_name"] == "depth"
         assert dataset.depth.attrs["long_name"] == "Depth"
+
+    def test_error_Coord_out_of_dataset_bounds(self):
+        try:
+            _ = subset(
+                dataset_id="cmems_mod_glo_phy_anfc_0.083deg_P1D-m",
+                start_datetime=datetime.today() + timedelta(10),
+                force_download=True,
+                end_datetime=datetime.today()
+                + timedelta(days=10, hours=23, minutes=59),
+            )
+        except core_functions.exceptions.CoordinatesOutOfDatasetBounds as e:
+            assert "Some or all of your subset selection" in e.__str__()
